refactor(inference): split BackgroundDownloadManager into focused files
Extract DownloadFileValidator (GGUF magic-byte and MLX directory checks)
into its own type with no instance-state dependencies, and move the
URLSessionDownloadDelegate conformance into a dedicated extension file.
BackgroundDownloadManager retains the session lifecycle, snapshot
coordination, disk-space check, persistence, and public API — down from
821 lines to ~600.
